Refactor: Move country and sector detection to utils, remove ultimate_main
