+++ PROFESSIONAL_SYSTEM_SUMMARY.md
@@ -105,15 +105,14 @@ You now have the **most professional, enterprise-grade healthcare discovery syst
 
 ```
 🏥 PROFESSIONAL HEALTHCARE DISCOVERY SYSTEM
-├── 🚀 professional_main.py       # ENTERPRISE MAIN SYSTEM
-├── 🎯 ultimate_main.py           # Ultimate version (backup)
-├── 🔍 ultimate_discoverer.py     # Advanced discovery engine
-├── ⚙️ ultimate_config.py         # 100+ database sources
-├── ✅ url_validator.py           # Professional validation
-├── 🛠️ utils.py                  # Utility functions
+├── 🚀 professional_main.py       # ENTERPRISE MAIN SYSTEM (43KB)
+├── 🔍 ultimate_discoverer.py     # Advanced discovery engine (22KB)
+├── ⚙️ ultimate_config.py         # 100+ database sources (16KB)
+├── ✅ url_validator.py           # Professional validation (14KB)
+├── 🛠️ utils.py                  # Enhanced utility functions (13KB)
 ├── 📋 requirements.txt           # Professional dependencies
-├── 📖 README.md                  # Comprehensive documentation
-└── 🏆 PROFESSIONAL_SYSTEM_SUMMARY.md  # This file
+├── 📖 README.md                  # Comprehensive documentation (15KB)
+└── 🏆 PROFESSIONAL_SYSTEM_SUMMARY.md  # Complete system guide (12KB)
 ```
 
 ---
