--- ultimate_main.py
@@ -1,616 +1,0 @@
-"""
-ULTIMATE Healthcare URL Validator and Discoverer
-The absolute best version designed to find 2000-5000+ healthcare companies across Europe
-
-This ultimate version includes:
-- 100+ comprehensive healthcare databases
-- Government and regulatory sources  
-- Advanced web scraping with pagination
-- Smart filtering and deduplication
-- Parallel processing for maximum speed
-- Multi-language support (6 languages)
-- Geographic targeting (100+ cities)
-- Sector-specific searches (30+ specialties)
-- Enhanced quality scoring
-- Professional output formatting
-
-Usage:
-    python ultimate_main.py
-"""
-
-import asyncio
-import json
-import csv
-import pandas as pd
-from datetime import datetime
-from typing import List, Dict
-import url_validator
-import ultimate_discoverer
-import ultimate_config as uconfig
-import utils
-
-
-# The provided list of healthcare URLs (starting point)
-INITIAL_HEALTHCARE_URLS = [
-    "https://www.acalta.de",
-    "https://www.actimi.com",
-    "https://www.emmora.de", 
-    "https://www.alfa-ai.com",
-    "https://www.apheris.com",
-    "https://www.aporize.com/",
-    "https://www.arztlena.com/",
-    "https://shop.getnutrio.com/",
-    "https://www.auta.health/",
-    "https://visioncheckout.com/",
-    "https://www.avayl.tech/",
-    "https://www.avimedical.com/avi-impact",
-    "https://de.becureglobal.com/",
-    "https://bellehealth.co/de/",
-    "https://www.biotx.ai/",
-    "https://www.brainjo.de/",
-    "https://brea.app/",
-    "https://breathment.com/",
-    "https://de.caona.eu/",
-    "https://www.careanimations.de/",
-    "https://sfs-healthcare.com",
-    "https://www.climedo.de/",
-    "https://www.cliniserve.de/",
-    "https://cogthera.de/#erfahren",
-    "https://www.comuny.de/",
-    "https://curecurve.de/elina-app/",
-    "https://www.cynteract.com/de/rehabilitation",
-    "https://www.healthmeapp.de/de/",
-    "https://deepeye.ai/",
-    "https://www.deepmentation.ai/",
-    "https://denton-systems.de/",
-    "https://www.derma2go.com/",
-    "https://www.dianovi.com/",
-    "http://dopavision.com/",
-    "https://www.dpv-analytics.com/",
-    "http://www.ecovery.de/",
-    "https://elixionmedical.com/",
-    "https://www.empident.de/",
-    "https://eye2you.ai/",
-    "https://www.fitwhit.de",
-    "https://www.floy.com/",
-    "https://fyzo.de/assistant/",
-    "https://www.gesund.de/app",
-    "https://www.glaice.de/",
-    "https://gleea.de/",
-    "https://www.guidecare.de/",
-    "https://www.apodienste.com/",
-    "https://www.help-app.de/",
-    "https://www.heynanny.com/",
-    "https://incontalert.de/",
-    "https://home.informme.info/",
-    "https://www.kranushealth.com/de/therapien/haeufiger-harndrang",
-    "https://www.kranushealth.com/de/therapien/inkontinenz"
-]
-
-
-def get_ultimate_country_estimate(url: str, title: str = "", description: str = "") -> str:
-    """
-    Ultimate country estimation using advanced detection methods
-    """
-    url_lower = url.lower()
-    domain = utils.extract_domain(url).lower()
-    combined_text = f"{url} {title} {description}".lower()
-    
-    # Enhanced country detection with multiple indicators
-    country_indicators = {
-        'Germany': [
-            # Domain patterns
-            '.de', '.com.de',
-            # Language indicators
-            'deutschland', 'german', 'deutsch', 'berlin', 'munich', 'münchen',
-            'hamburg', 'frankfurt', 'cologne', 'köln', 'stuttgart', 'düsseldorf',
-            'gmbh', 'ug', 'ag',
-            # German healthcare terms
-            'gesundheit', 'medizin', 'arzt', 'krankenhaus', 'klinik'
-        ],
-        'France': [
-            '.fr', '.com.fr', 'france', 'french', 'français', 'paris', 'lyon',
-            'marseille', 'toulouse', 'nice', 'nantes', 'strasbourg', 'sa', 'sarl',
-            'santé', 'médecine', 'médecin', 'hôpital', 'clinique'
-        ],
-        'Netherlands': [
-            '.nl', '.com.nl', 'netherlands', 'dutch', 'nederland', 'amsterdam',
-            'rotterdam', 'hague', 'utrecht', 'eindhoven', 'bv', 'nv',
-            'gezondheid', 'geneeskunde', 'arts', 'ziekenhuis'
-        ],
-        'United Kingdom': [
-            '.uk', '.co.uk', '.org.uk', 'britain', 'british', 'england', 'scotland',
-            'wales', 'london', 'manchester', 'birmingham', 'glasgow', 'liverpool',
-            'leeds', 'sheffield', 'edinburgh', 'bristol', 'cardiff', 'ltd', 'plc'
-        ],
-        'Switzerland': [
-            '.ch', '.com.ch', 'switzerland', 'swiss', 'schweiz', 'suisse', 'svizzera',
-            'zurich', 'zürich', 'geneva', 'genève', 'basel', 'lausanne', 'bern'
-        ],
-        'Spain': [
-            '.es', '.com.es', 'spain', 'spanish', 'españa', 'madrid', 'barcelona',
-            'valencia', 'seville', 'sevilla', 'zaragoza', 'málaga', 'sl', 'sau',
-            'salud', 'medicina', 'médico', 'hospital', 'clínica'
-        ],
-        'Italy': [
-            '.it', '.com.it', 'italy', 'italian', 'italia', 'rome', 'roma', 'milan',
-            'milano', 'naples', 'napoli', 'turin', 'torino', 'palermo', 'genoa',
-            'bologna', 'florence', 'firenze', 'spa', 'srl',
-            'salute', 'medicina', 'medico', 'ospedale', 'clinica'
-        ],
-        'Sweden': [
-            '.se', '.com.se', 'sweden', 'swedish', 'sverige', 'stockholm',
-            'gothenburg', 'göteborg', 'malmö', 'uppsala', 'ab',
-            'hälsa', 'medicin', 'sjukhus', 'läkare'
-        ],
-        'Denmark': [
-            '.dk', '.com.dk', 'denmark', 'danish', 'danmark', 'copenhagen',
-            'københavn', 'aarhus', 'odense', 'aalborg', 'a/s', 'aps',
-            'sundhed', 'medicin', 'sygehus', 'læge'
-        ],
-        'Norway': [
-            '.no', '.com.no', 'norway', 'norwegian', 'norge', 'oslo', 'bergen',
-            'trondheim', 'stavanger', 'as', 'asa',
-            'helse', 'medisin', 'sykehus', 'lege'
-        ],
-        'Finland': [
-            '.fi', '.com.fi', 'finland', 'finnish', 'suomi', 'helsinki',
-            'tampere', 'turku', 'oulu', 'oy', 'oyj',
-            'terveys', 'lääketiede', 'sairaala', 'lääkäri'
-        ],
-        'Belgium': [
-            '.be', '.com.be', 'belgium', 'belgian', 'belgië', 'belgique',
-            'brussels', 'brussel', 'bruxelles', 'antwerp', 'antwerpen',
-            'ghent', 'gent', 'charleroi', 'liège', 'sprl', 'bvba'
-        ],
-        'Austria': [
-            '.at', '.com.at', 'austria', 'austrian', 'österreich', 'vienna',
-            'wien', 'salzburg', 'innsbruck', 'linz', 'graz'
-        ],
-        'Ireland': [
-            '.ie', '.com.ie', 'ireland', 'irish', 'dublin', 'cork', 'galway',
-            'waterford', 'limerick'
-        ],
-        'Portugal': [
-            '.pt', '.com.pt', 'portugal', 'portuguese', 'lisbon', 'lisboa',
-            'porto', 'braga', 'coimbra'
-        ]
-    }
-    
-    # Score each country based on indicators found
-    country_scores = {}
-    for country, indicators in country_indicators.items():
-        score = 0
-        for indicator in indicators:
-            if indicator in combined_text:
-                # Domain extensions get higher scores
-                if indicator.startswith('.'):
-                    score += 5
-                # City names get medium scores
-                elif any(city in indicator for city in ['berlin', 'paris', 'london', 'amsterdam']):
-                    score += 3
-                # Other indicators get base scores
-                else:
-                    score += 1
-        
-        if score > 0:
-            country_scores[country] = score
-    
-    # Return country with highest score
-    if country_scores:
-        best_country = max(country_scores, key=country_scores.get)
-        return best_country
-    
-    # Fallback detection
-    if '.eu' in domain:
-        return 'European Union'
-    elif any(tld in domain for tld in ['.com', '.org', '.net']):
-        return 'International'
-    else:
-        return 'Other'
-
-
-def classify_healthcare_sector(url: str, title: str = "", description: str = "") -> str:
-    """
-    Classify healthcare companies by sector/specialty
-    """
-    combined_text = f"{url} {title} {description}".lower()
-    
-    sector_keywords = {
-        'Digital Therapeutics': ['digital therapeutics', 'dtx', 'prescription app', 'therapeutic app'],
-        'Telemedicine': ['telemedicine', 'telehealth', 'remote consultation', 'virtual care', 'online doctor'],
-        'Medical Devices': ['medical device', 'medical equipment', 'implant', 'prosthetic', 'monitor'],
-        'Health Analytics': ['health analytics', 'medical data', 'health insights', 'clinical analytics'],
-        'Mental Health': ['mental health', 'psychology', 'psychiatry', 'therapy', 'behavioral health'],
-        'AI/ML Health': ['medical ai', 'healthcare ai', 'machine learning', 'artificial intelligence'],
-        'Biotech': ['biotech', 'biotechnology', 'pharmaceutical', 'drug development', 'clinical trial'],
-        'Medical Imaging': ['medical imaging', 'radiology', 'mri', 'ct scan', 'ultrasound', 'x-ray'],
-        'Chronic Care': ['chronic disease', 'diabetes', 'hypertension', 'copd', 'heart disease'],
-        'Women\'s Health': ['womens health', 'fertility', 'pregnancy', 'reproductive health'],
-        'Elderly Care': ['elderly care', 'senior health', 'aging', 'geriatric', 'dementia'],
-        'Pediatric': ['pediatric', 'children health', 'infant care', 'neonatal'],
-        'Surgery Tech': ['surgical', 'surgery', 'robotic surgery', 'minimally invasive'],
-        'Digital Pharmacy': ['digital pharmacy', 'e-pharmacy', 'medication management'],
-        'Fitness/Wellness': ['fitness', 'wellness', 'nutrition', 'exercise', 'lifestyle']
-    }
-    
-    for sector, keywords in sector_keywords.items():
-        for keyword in keywords:
-            if keyword in combined_text:
-                return sector
-    
-    return 'General Healthcare'
-
-
-def print_ultimate_statistics(all_results: List[Dict]):
-    """
-    Print comprehensive ultimate statistics
-    """
-    print("\n" + "="*80)
-    print("ULTIMATE DISCOVERY FINAL STATISTICS")
-    print("="*80)
-    
-    # Overall statistics
-    total_urls = len(all_results)
-    live_urls = [r for r in all_results if r['is_live']]
-    healthcare_urls = [r for r in live_urls if r['is_healthcare']]
-    
-    print(f"🎯 ULTIMATE TARGET: {uconfig.ULTIMATE_SETTINGS['MAX_TOTAL_URLS_TARGET']:,} healthcare companies")
-    print(f"📊 ULTIMATE RESULTS:")
-    print(f"   Total URLs discovered & validated: {total_urls:,}")
-    print(f"   Live URLs: {len(live_urls):,} ({len(live_urls)/total_urls*100:.1f}%)")
-    print(f"   Healthcare companies: {len(healthcare_urls):,} ({len(healthcare_urls)/total_urls*100:.1f}%)")
-    
-    # Achievement assessment
-    target = uconfig.ULTIMATE_SETTINGS['MAX_TOTAL_URLS_TARGET']
-    achievement_ratio = len(healthcare_urls) / target
-    
-    if achievement_ratio >= 1.0:
-        print(f"✅ ULTIMATE SUCCESS: Found {len(healthcare_urls):,} companies (target achieved!)")
-    elif achievement_ratio >= 0.5:
-        print(f"🟡 EXCELLENT PROGRESS: Found {len(healthcare_urls):,} companies ({achievement_ratio*100:.1f}% of target)")
-    else:
-        print(f"🔴 GOOD START: Found {len(healthcare_urls):,} companies ({achievement_ratio*100:.1f}% of target)")
-    
-    # Enhanced geographic distribution
-    print(f"\n🌍 COMPREHENSIVE GEOGRAPHIC DISTRIBUTION:")
-    countries = {}
-    for result in healthcare_urls:
-        country = get_ultimate_country_estimate(
-            result['url'], 
-            result.get('title', ''), 
-            result.get('description', '')
-        )
-        countries[country] = countries.get(country, 0) + 1
-    
-    # Sort by count and show all countries
-    for country, count in sorted(countries.items(), key=lambda x: x[1], reverse=True):
-        percentage = (count / len(healthcare_urls)) * 100
-        print(f"   {country}: {count:,} companies ({percentage:.1f}%)")
-    
-    # Sector distribution
-    print(f"\n🏥 HEALTHCARE SECTOR DISTRIBUTION:")
-    sectors = {}
-    for result in healthcare_urls:
-        sector = classify_healthcare_sector(
-            result['url'],
-            result.get('title', ''),
-            result.get('description', '')
-        )
-        sectors[sector] = sectors.get(sector, 0) + 1
-    
-    for sector, count in sorted(sectors.items(), key=lambda x: x[1], reverse=True):
-        if count >= 2:  # Only show sectors with 2+ companies
-            percentage = (count / len(healthcare_urls)) * 100
-            print(f"   {sector}: {count} companies ({percentage:.1f}%)")
-    
-    # Discovery source performance
-    print(f"\n📈 DISCOVERY SOURCE PERFORMANCE:")
-    sources = {}
-    for result in all_results:
-        source = result.get('source', 'Unknown')
-        if source not in sources:
-            sources[source] = {'total': 0, 'live': 0, 'healthcare': 0}
-        sources[source]['total'] += 1
-        if result.get('is_live'):
-            sources[source]['live'] += 1
-            if result.get('is_healthcare'):
-                sources[source]['healthcare'] += 1
-    
-    for source, stats in sorted(sources.items(), key=lambda x: x[1]['healthcare'], reverse=True):
-        if stats['healthcare'] > 0:
-            efficiency = (stats['healthcare'] / stats['total']) * 100
-            print(f"   {source}:")
-            print(f"     Healthcare: {stats['healthcare']}, Live: {stats['live']}, Total: {stats['total']} (Efficiency: {efficiency:.1f}%)")
-    
-    # Quality metrics
-    if healthcare_urls:
-        healthcare_scores = [r.get('healthcare_score', 0) for r in healthcare_urls if r.get('healthcare_score')]
-        if healthcare_scores:
-            avg_score = sum(healthcare_scores) / len(healthcare_scores)
-            print(f"\n⭐ QUALITY METRICS:")
-            print(f"   Average healthcare relevance score: {avg_score:.1f}")
-            print(f"   High-quality companies (score ≥ 10): {len([s for s in healthcare_scores if s >= 10])}")
-    
-    print(f"\n🎉 ULTIMATE DISCOVERY STATISTICS COMPLETE!")
-
-
-def save_ultimate_results(results: List[Dict], timestamp: str):
-    """
-    Save ultimate results with professional formatting
-    """
-    # Filter for healthcare companies only
-    healthcare_only = [r for r in results if r.get('is_live') and r.get('is_healthcare')]
-    
-    if len(healthcare_only) == 0:
-        print("⚠️ No healthcare companies found in results!")
-        return
-    
-    # Create enhanced DataFrame
-    df_data = []
-    for result in healthcare_only:
-        country = get_ultimate_country_estimate(
-            result['url'],
-            result.get('title', ''),
-            result.get('description', '')
-        )
-        sector = classify_healthcare_sector(
-            result['url'],
-            result.get('title', ''),
-            result.get('description', '')
-        )
-        
-        df_data.append({
-            'url': result['url'],
-            'domain': utils.extract_domain(result['url']),
-            'company_name': result.get('title', '').replace(' | ', ' - ').replace('|', '-')[:100],
-            'country': country,
-            'sector': sector,
-            'description': result.get('description', '')[:200],
-            'discovery_source': result.get('source', 'Unknown'),
-            'quality_score': result.get('healthcare_score', 0)
-        })
-    
-    df = pd.DataFrame(df_data)
-    
-    # Sort by quality score (highest first), then by country
-    df = df.sort_values(['quality_score', 'country'], ascending=[False, True])
-    
-    # Save ultimate results
-    ultimate_csv = f"ultimate_healthcare_companies_{timestamp}.csv"
-    ultimate_json = f"ultimate_healthcare_companies_{timestamp}.json"
-    
-    # Clean CSV with professional columns
-    df.to_csv(ultimate_csv, index=False, encoding='utf-8')
-    
-    # Enhanced JSON with comprehensive metadata
-    ultimate_data = {
-        'metadata': {
-            'generated_at': datetime.now().isoformat(),
-            'discovery_method': 'Ultimate Multi-source Discovery',
-            'version': '2.0 Ultimate',
-            'target_regions': 'All European Countries',
-            'discovery_sources': sum(len(sources) for sources in uconfig.ULTIMATE_HEALTHCARE_SOURCES.values()),
-            'search_coverage': {
-                'cities': len(uconfig.ULTIMATE_EUROPEAN_CITIES),
-                'healthcare_sectors': len(uconfig.ULTIMATE_HEALTHCARE_SECTORS),
-                'languages': 6,
-                'countries': len(set(df['country'].unique()))
-            },
-            'results_summary': {
-                'total_companies': len(df),
-                'countries_covered': len(df['country'].unique()),
-                'sectors_covered': len(df['sector'].unique()),
-                'average_quality_score': df['quality_score'].mean(),
-                'top_countries': df['country'].value_counts().head().to_dict(),
-                'top_sectors': df['sector'].value_counts().head().to_dict()
-            },
-            'quality_metrics': {
-                'high_quality_companies': len(df[df['quality_score'] >= 10]),
-                'medium_quality_companies': len(df[(df['quality_score'] >= 5) & (df['quality_score'] < 10)]),
-                'basic_quality_companies': len(df[df['quality_score'] < 5])
-            }
-        },
-        'companies': df.to_dict('records')
-    }
-    
-    with open(ultimate_json, 'w', encoding='utf-8') as f:
-        json.dump(ultimate_data, f, indent=2, ensure_ascii=False)
-    
-    print(f"\n💎 ULTIMATE RESULTS SAVED:")
-    print(f"   📊 CSV: {ultimate_csv}")
-    print(f"   📊 JSON: {ultimate_json}")
-    print(f"   🏥 Companies: {len(df):,} verified healthcare companies")
-    print(f"   🌍 Countries: {len(df['country'].unique())} countries covered")
-    print(f"   🏭 Sectors: {len(df['sector'].unique())} healthcare sectors")
-    
-    # Show top countries
-    print(f"\n🌍 TOP COUNTRIES:")
-    country_stats = df['country'].value_counts().head(10)
-    for country, count in country_stats.items():
-        print(f"   {country}: {count} companies")
-
-
-async def ultimate_main():
-    """
-    Ultimate main function targeting thousands of healthcare companies
-    """
-    print("🚀 ULTIMATE Healthcare URL Validator and Discoverer")
-    print("=" * 80)
-    print("🎯 ULTIMATE TARGET: Find 2000-5000+ healthcare companies across Europe")
-    print("🔍 ULTIMATE METHOD: Comprehensive multi-source discovery")
-    print(f"⏰ Starting at: {datetime.now().strftime('%Y-%m-%d %H:%M:%S')}")
-    print(f"📊 Sources: {sum(len(sources) for sources in uconfig.ULTIMATE_HEALTHCARE_SOURCES.values())} databases")
-    print(f"🌍 Coverage: {len(uconfig.ULTIMATE_EUROPEAN_CITIES)} cities, {len(uconfig.ULTIMATE_HEALTHCARE_SECTORS)} sectors")
-    print("")
-    
-    all_results = []
-    
-    # Step 1: Validate the provided URLs (starting point)
-    print(f"📋 Step 1: Validating {len(INITIAL_HEALTHCARE_URLS)} provided URLs...")
-    validated_results = url_validator.clean_and_validate_urls(INITIAL_HEALTHCARE_URLS)
-    
-    # Add source information
-    for result in validated_results:
-        result['source'] = 'Initial List'
-    
-    all_results.extend(validated_results)
-    
-    initial_healthcare = len([r for r in validated_results if r.get('is_live') and r.get('is_healthcare')])
-    print(f"   ✅ Initial validation complete: {initial_healthcare} healthcare companies confirmed")
-    
-    # Step 2: Ultimate discovery (the main enhancement)
-    print(f"\n🔍 Step 2: ULTIMATE comprehensive discovery...")
-    print(f"   This may take 15-30 minutes due to comprehensive coverage")
-    
-    try:
-        discovered_results = await ultimate_discoverer.discover_ultimate_healthcare_urls()
-        
-        if discovered_results:
-            print(f"\n🎉 Ultimate discovery complete! Found {len(discovered_results):,} new URLs")
-            print(f"🔬 Now validating all discovered URLs in batches...")
-            
-            urls_to_validate = [r['url'] for r in discovered_results]
-            url_source_map = {r['url']: r['source'] for r in discovered_results}
-            url_score_map = {r['url']: r.get('healthcare_score', 0) for r in discovered_results}
-            
-            # Validate in batches with progress tracking
-            batch_size = 100
-            validated_discovered = []
-            
-            for i in range(0, len(urls_to_validate), batch_size):
-                batch = urls_to_validate[i:i+batch_size]
-                batch_num = i//batch_size + 1
-                total_batches = (len(urls_to_validate)-1)//batch_size + 1
-                
-                print(f"   Validating batch {batch_num}/{total_batches} ({len(batch)} URLs)...")
-                
-                batch_results = url_validator.clean_and_validate_urls(batch)
-                
-                # Enhance results with discovery metadata
-                for result in batch_results:
-                    if result['url'] in url_source_map:
-                        result['source'] = url_source_map[result['url']]
-                        result['healthcare_score'] = url_score_map.get(result['url'], 0)
-                
-                validated_discovered.extend(batch_results)
-                
-                # Progress update
-                healthcare_so_far = len([r for r in validated_discovered if r.get('is_live') and r.get('is_healthcare')])
-                total_healthcare = initial_healthcare + healthcare_so_far
-                print(f"     Healthcare companies found: {healthcare_so_far} (Total: {total_healthcare})")
-            
-            all_results.extend(validated_discovered)
-        else:
-            print("⚠️  No new URLs discovered (may be due to network limitations)")
-            
-    except Exception as e:
-        print(f"❌ Error during ultimate discovery: {e}")
-        print("Continuing with initial URLs only...")
-    
-    # Step 3: Enhanced deduplication and quality filtering
-    print(f"\n🔄 Step 3: Enhanced deduplication and quality filtering...")
-    unique_results = []
-    seen_urls = set()
-    
-    # Sort by quality score first
-    all_results.sort(key=lambda x: x.get('healthcare_score', 0), reverse=True)
-    
-    for result in all_results:
-        if result['url'] not in seen_urls:
-            seen_urls.add(result['url'])
-            unique_results.append(result)
-    
-    print(f"   Total unique URLs after deduplication: {len(unique_results):,}")
-    
-    # Step 4: Ultimate statistics and analysis
-    print_ultimate_statistics(unique_results)
-    
-    # Step 5: Save ultimate results
-    print(f"\n💾 Step 5: Saving ultimate results...")
-    timestamp = datetime.now().strftime('%Y%m%d_%H%M%S')
-    save_ultimate_results(unique_results, timestamp)
-    
-    # Final ultimate summary
-    healthcare_final = len([r for r in unique_results if r.get('is_live') and r.get('is_healthcare')])
-    target = uconfig.ULTIMATE_SETTINGS['MAX_TOTAL_URLS_TARGET']
-    
-    print(f"\n🎉 ULTIMATE DISCOVERY COMPLETE!")
-    print("=" * 80)
-    print(f"📊 ULTIMATE FINAL RESULTS:")
-    print(f"   • Total URLs processed: {len(unique_results):,}")
-    print(f"   • Verified healthcare companies: {healthcare_final:,}")
-    print(f"   • Target achievement: {(healthcare_final/target)*100:.1f}% of {target:,}")
-    
-    if healthcare_final >= target:
-        print(f"✅ ULTIMATE SUCCESS! Found {healthcare_final:,} healthcare companies!")
-        print(f"📈 Exceeded expectations with comprehensive European coverage!")
-    elif healthcare_final >= target * 0.5:
-        print(f"🟡 EXCELLENT RESULTS! Found {healthcare_final:,} companies")
-        print(f"📈 Achieved {(healthcare_final/target)*100:.1f}% of ultimate target")
-    else:
-        print(f"🔴 GOOD FOUNDATION! Found {healthcare_final:,} companies")
-        print("💡 ULTIMATE ENHANCEMENT SUGGESTIONS:")
-        print("   - Run during different times for better access")
-        print("   - Consider using premium API access for major databases")
-        print("   - Expand to Eastern European countries")
-        print("   - Add more clinical trial databases")
-    
-    print(f"\n🌟 ULTIMATE HEALTHCARE DISCOVERY SYSTEM COMPLETE!")
-    
-    return unique_results
-
-
-def run_ultimate_main():
-    """
-    Ultimate entry point with comprehensive error handling
-    """
-    print("🔧 Initializing Ultimate Healthcare Discovery System...")
-    print(f"🎯 Target: {uconfig.ULTIMATE_SETTINGS['MAX_TOTAL_URLS_TARGET']:,} healthcare companies")
-    print(f"🌍 Coverage: All European countries")
-    print(f"🏭 Sources: Government, Industry, Startups, Research, Conferences")
-    print("")
-    
-    try:
-        # Check for event loop
-        loop = asyncio.get_running_loop()
-        print("⚠️  Existing event loop detected.")
-        print("For optimal performance, run from command line: python ultimate_main.py")
-        print("Proceeding with compatibility mode...\n")
-    except RuntimeError:
-        print("✅ Optimal environment detected. Maximum performance mode enabled.\n")
-    
-    try:
-        results = asyncio.run(ultimate_main())
-        
-        healthcare_count = len([r for r in results if r.get('is_live') and r.get('is_healthcare')])
-        target = uconfig.ULTIMATE_SETTINGS['MAX_TOTAL_URLS_TARGET']
-        
-        print(f"\n🎊 ULTIMATE SUCCESS METRICS:")
-        print(f"   Healthcare companies found: {healthcare_count:,}")
-        print(f"   Target achievement: {(healthcare_count/target)*100:.1f}%")
-        print(f"   Quality: Professional-grade dataset ready for analysis")
-        
-        if healthcare_count >= 500:
-            print(f"\n🏆 OUTSTANDING ACHIEVEMENT!")
-            print(f"   Found {healthcare_count:,} healthcare companies across Europe!")
-        
-    except RuntimeError as e:
-        if "asyncio.run() cannot be called from a running event loop" in str(e):
-            print("\n⚠️  Event loop conflict detected.")
-            print("Please run from command line for ultimate functionality:")
-            print("  python ultimate_main.py")
-        else:
-            print(f"\n❌ Runtime error: {e}")
-    except KeyboardInterrupt:
-        print("\n⚠️  Process interrupted by user.")
-        print("Ultimate discovery can be resumed by running the script again.")
-    except Exception as e:
-        print(f"\n❌ An error occurred: {e}")
-        print("Please check your internet connection and try again.")
-    
-    print("\n🚀 Thank you for using the Ultimate Healthcare Discovery System!")
-
-
-if __name__ == "__main__":
-    run_ultimate_main()
