+++ utils.py
@@ -7,7 +7,26 @@ import re
 import tldextract
 from urllib.parse import urlparse, urljoin
 from typing import List, Set, Optional
-import config
+import ultimate_config as config
+
+# Healthcare keywords for content detection
+HEALTHCARE_KEYWORDS = [
+    'health', 'medical', 'healthcare', 'medicine', 'clinic', 'hospital',
+    'doctor', 'physician', 'patient', 'therapy', 'treatment', 'diagnosis',
+    'pharmaceutical', 'biotech', 'medtech', 'digital health', 'telemedicine'
+]
+
+# Domains to exclude
+EXCLUDED_DOMAINS = [
+    'google.com', 'facebook.com', 'linkedin.com', 'twitter.com', 'instagram.com',
+    'youtube.com', 'wikipedia.org', 'crunchbase.com', 'angel.co'
+]
+
+# URL patterns to exclude
+EXCLUDED_PATTERNS = [
+    '/login', '/signin', '/signup', '/register', '/auth', '/oauth',
+    '/privacy', '/terms', '/contact', '/about/team', '/careers'
+]
 
 
 def clean_url(url: str) -> str:
@@ -56,7 +75,7 @@ def is_healthcare_related(text: str, url: str = "") -> bool:
     
     # Check for healthcare keywords
     healthcare_score = 0
-    for keyword in config.HEALTHCARE_KEYWORDS:
+    for keyword in HEALTHCARE_KEYWORDS:
         if keyword.lower() in combined_text:
             healthcare_score += 1
     
@@ -80,12 +99,12 @@ def should_exclude_url(url: str) -> bool:
         path = parsed.path.lower()
         
         # Check excluded domains
-        for excluded_domain in config.EXCLUDED_DOMAINS:
+        for excluded_domain in EXCLUDED_DOMAINS:
             if excluded_domain in domain:
                 return True
         
         # Check excluded path patterns
-        for pattern in config.EXCLUDED_PATTERNS:
+        for pattern in EXCLUDED_PATTERNS:
             if pattern in path:
                 return True
                 
@@ -219,3 +238,157 @@ def normalize_text(text: str) -> str:
     normalized = re.sub(r'[^\w\s.,;:!?-]', ' ', normalized)
     
     return normalized
+
+
+def get_ultimate_country_estimate(url: str, title: str = "", description: str = "") -> str:
+    """
+    Ultimate country estimation using advanced detection methods
+    """
+    url_lower = url.lower()
+    domain = extract_domain(url).lower()
+    combined_text = f"{url} {title} {description}".lower()
+    
+    # Enhanced country detection with multiple indicators
+    country_indicators = {
+        'Germany': [
+            # Domain patterns
+            '.de', '.com.de',
+            # Language indicators
+            'deutschland', 'german', 'deutsch', 'berlin', 'munich', 'münchen',
+            'hamburg', 'frankfurt', 'cologne', 'köln', 'stuttgart', 'düsseldorf',
+            'gmbh', 'ug', 'ag',
+            # German healthcare terms
+            'gesundheit', 'medizin', 'arzt', 'krankenhaus', 'klinik'
+        ],
+        'France': [
+            '.fr', '.com.fr', 'france', 'french', 'français', 'paris', 'lyon',
+            'marseille', 'toulouse', 'nice', 'nantes', 'strasbourg', 'sa', 'sarl',
+            'santé', 'médecine', 'médecin', 'hôpital', 'clinique'
+        ],
+        'Netherlands': [
+            '.nl', '.com.nl', 'netherlands', 'dutch', 'nederland', 'amsterdam',
+            'rotterdam', 'hague', 'utrecht', 'eindhoven', 'bv', 'nv',
+            'gezondheid', 'geneeskunde', 'arts', 'ziekenhuis'
+        ],
+        'United Kingdom': [
+            '.uk', '.co.uk', '.org.uk', 'britain', 'british', 'england', 'scotland',
+            'wales', 'london', 'manchester', 'birmingham', 'glasgow', 'liverpool',
+            'leeds', 'sheffield', 'edinburgh', 'bristol', 'cardiff', 'ltd', 'plc'
+        ],
+        'Switzerland': [
+            '.ch', '.com.ch', 'switzerland', 'swiss', 'schweiz', 'suisse', 'svizzera',
+            'zurich', 'zürich', 'geneva', 'genève', 'basel', 'lausanne', 'bern'
+        ],
+        'Spain': [
+            '.es', '.com.es', 'spain', 'spanish', 'españa', 'madrid', 'barcelona',
+            'valencia', 'seville', 'sevilla', 'zaragoza', 'málaga', 'sl', 'sau',
+            'salud', 'medicina', 'médico', 'hospital', 'clínica'
+        ],
+        'Italy': [
+            '.it', '.com.it', 'italy', 'italian', 'italia', 'rome', 'roma', 'milan',
+            'milano', 'naples', 'napoli', 'turin', 'torino', 'palermo', 'genoa',
+            'bologna', 'florence', 'firenze', 'spa', 'srl',
+            'salute', 'medicina', 'medico', 'ospedale', 'clinica'
+        ],
+        'Sweden': [
+            '.se', '.com.se', 'sweden', 'swedish', 'sverige', 'stockholm',
+            'gothenburg', 'göteborg', 'malmö', 'uppsala', 'ab',
+            'hälsa', 'medicin', 'sjukhus', 'läkare'
+        ],
+        'Denmark': [
+            '.dk', '.com.dk', 'denmark', 'danish', 'danmark', 'copenhagen',
+            'københavn', 'aarhus', 'odense', 'aalborg', 'a/s', 'aps',
+            'sundhed', 'medicin', 'sygehus', 'læge'
+        ],
+        'Norway': [
+            '.no', '.com.no', 'norway', 'norwegian', 'norge', 'oslo', 'bergen',
+            'trondheim', 'stavanger', 'as', 'asa',
+            'helse', 'medisin', 'sykehus', 'lege'
+        ],
+        'Finland': [
+            '.fi', '.com.fi', 'finland', 'finnish', 'suomi', 'helsinki',
+            'tampere', 'turku', 'oulu', 'oy', 'oyj',
+            'terveys', 'lääketiede', 'sairaala', 'lääkäri'
+        ],
+        'Belgium': [
+            '.be', '.com.be', 'belgium', 'belgian', 'belgië', 'belgique',
+            'brussels', 'brussel', 'bruxelles', 'antwerp', 'antwerpen',
+            'ghent', 'gent', 'charleroi', 'liège', 'sprl', 'bvba'
+        ],
+        'Austria': [
+            '.at', '.com.at', 'austria', 'austrian', 'österreich', 'vienna',
+            'wien', 'salzburg', 'innsbruck', 'linz', 'graz'
+        ],
+        'Ireland': [
+            '.ie', '.com.ie', 'ireland', 'irish', 'dublin', 'cork', 'galway',
+            'waterford', 'limerick'
+        ],
+        'Portugal': [
+            '.pt', '.com.pt', 'portugal', 'portuguese', 'lisbon', 'lisboa',
+            'porto', 'braga', 'coimbra'
+        ]
+    }
+    
+    # Score each country based on indicators found
+    country_scores = {}
+    for country, indicators in country_indicators.items():
+        score = 0
+        for indicator in indicators:
+            if indicator in combined_text:
+                # Domain extensions get higher scores
+                if indicator.startswith('.'):
+                    score += 5
+                # City names get medium scores
+                elif any(city in indicator for city in ['berlin', 'paris', 'london', 'amsterdam']):
+                    score += 3
+                # Other indicators get base scores
+                else:
+                    score += 1
+        
+        if score > 0:
+            country_scores[country] = score
+    
+    # Return country with highest score
+    if country_scores:
+        best_country = max(country_scores, key=country_scores.get)
+        return best_country
+    
+    # Fallback detection
+    if '.eu' in domain:
+        return 'European Union'
+    elif any(tld in domain for tld in ['.com', '.org', '.net']):
+        return 'International'
+    else:
+        return 'Other'
+
+
+def classify_healthcare_sector(url: str, title: str = "", description: str = "") -> str:
+    """
+    Classify healthcare companies by sector/specialty
+    """
+    combined_text = f"{url} {title} {description}".lower()
+    
+    sector_keywords = {
+        'Digital Therapeutics': ['digital therapeutics', 'dtx', 'prescription app', 'therapeutic app'],
+        'Telemedicine': ['telemedicine', 'telehealth', 'remote consultation', 'virtual care', 'online doctor'],
+        'Medical Devices': ['medical device', 'medical equipment', 'implant', 'prosthetic', 'monitor'],
+        'Health Analytics': ['health analytics', 'medical data', 'health insights', 'clinical analytics'],
+        'Mental Health': ['mental health', 'psychology', 'psychiatry', 'therapy', 'behavioral health'],
+        'AI/ML Health': ['medical ai', 'healthcare ai', 'machine learning', 'artificial intelligence'],
+        'Biotech': ['biotech', 'biotechnology', 'pharmaceutical', 'drug development', 'clinical trial'],
+        'Medical Imaging': ['medical imaging', 'radiology', 'mri', 'ct scan', 'ultrasound', 'x-ray'],
+        'Chronic Care': ['chronic disease', 'diabetes', 'hypertension', 'copd', 'heart disease'],
+        'Women\'s Health': ['womens health', 'fertility', 'pregnancy', 'reproductive health'],
+        'Elderly Care': ['elderly care', 'senior health', 'aging', 'geriatric', 'dementia'],
+        'Pediatric': ['pediatric', 'children health', 'infant care', 'neonatal'],
+        'Surgery Tech': ['surgical', 'surgery', 'robotic surgery', 'minimally invasive'],
+        'Digital Pharmacy': ['digital pharmacy', 'e-pharmacy', 'medication management'],
+        'Fitness/Wellness': ['fitness', 'wellness', 'nutrition', 'exercise', 'lifestyle']
+    }
+    
+    for sector, keywords in sector_keywords.items():
+        for keyword in keywords:
+            if keyword in combined_text:
+                return sector
+    
+    return 'General Healthcare'
